Swift: generalize open redirection on both platforms and on outputs
The hash map mechanism that was already in use for reading swiftmodule
files on macOS is now in use also on renames and on Linux. This replaces
the whole output flag rewrite mechanism.

Note: all writes will also be redirected in order to avoid our extractor
changing the output files of the original build, but for the moment
in order to function only intercepting renames of `.swiftmodule` files
is needed, as swift creates temporaries and moves them into place when
finished writing.